@@ -1,75 +1,72 @@
 {
   "name": "material-design-lite",
-  "version": "1.0.6",
+  "version": "1.1.3",
   "description": "Material Design Components in CSS, JS and HTML",
   "private": true,
   "license": "Apache-2.0",
   "author": "Google",
   "repository": "google/material-design-lite",
+  "main": "dist/material.min.js",
   "devDependencies": {
     "acorn": "^2.2.0",
-    "apache-server-configs": "^2.7.1",
+    "babel-core": "^5.8.25",
     "browser-sync": "^2.2.3",
-    "chai": "^2.0.0",
+    "chai": "^3.3.0",
     "chai-jquery": "^2.0.0",
-    "del": "^1.1.1",
-    "drool": "^0.2.1",
+    "del": "^2.0.2",
+    "drool": "^0.3.1",
     "escodegen": "^1.6.1",
-    "gulp": "^3.8.11",
-    "gulp-autoprefixer": "^2.0.0",
-    "gulp-cache": "^0.2.6",
-    "gulp-changed": "^1.0.0",
+    "google-closure-compiler": "",
+    "gulp": "^3.9.0",
+    "gulp-autoprefixer": "^3.0.2",
+    "gulp-cache": "^0.3.0",
+    "gulp-closure-compiler": "^0.3.1",
     "gulp-concat": "^2.4.1",
     "gulp-connect": "^2.2.0",
     "gulp-css-inline-images": "^0.1.1",
-    "gulp-csso": "^1.0.0",
+    "gulp-csso": "1.0.0",
     "gulp-file": "^0.2.0",
-    "gulp-filter": "^2.0.2",
-    "gulp-flatten": "0.0.4",
+    "gulp-flatten": "^0.2.0",
     "gulp-front-matter": "^1.2.2",
-    "gulp-gh-pages": "^0.5.0",
     "gulp-header": "^1.2.2",
-    "gulp-if": "^1.2.1",
-    "gulp-iife": "0.0.7",
+    "gulp-if": "^2.0.0",
+    "gulp-iife": "^0.1.0",
     "gulp-imagemin": "^2.2.1",
-    "gulp-jscs": "^3.0.0",
+    "gulp-jscs": "^3.0.1",
     "gulp-jshint": "^1.6.3",
     "gulp-load-plugins": "^0.10.0",
-    "gulp-markdown": "^1.0.0",
     "gulp-marked": "^1.0.0",
-    "gulp-minify-html": "^1.0.0",
-    "gulp-mocha-phantomjs": "^0.6.1",
-    "gulp-open": "^0.3.2",
+    "gulp-mocha-phantomjs": "^0.10.1",
+    "gulp-open": "^1.0.0",
     "gulp-rename": "^1.2.0",
     "gulp-replace": "^0.5.3",
     "gulp-sass": "2.0.*",
     "gulp-shell": "^0.4.2",
-    "gulp-size": "^1.0.0",
+    "gulp-size": "^2.0.0",
     "gulp-sourcemaps": "^1.3.0",
     "gulp-subtree": "^0.1.0",
     "gulp-tap": "^0.1.3",
     "gulp-uglify": "^1.0.1",
-    "gulp-useref": "^1.0.1",
     "gulp-util": "^3.0.4",
     "gulp-zip": "^3.0.2",
     "humanize": "0.0.9",
     "jquery": "^2.1.3",
-    "jshint-stylish": "^1.0.0",
-    "merge-stream": "^0.1.7",
+    "jshint-stylish": "^2.0.1",
+    "merge-stream": "^1.0.0",
     "mocha": "^2.1.0",
-    "opn": "^1.0.0",
-    "phantomjs": "^1.9.17",
     "prismjs": "0.0.1",
-    "require-dir": "^0.3.0",
     "run-sequence": "^1.0.2",
     "swig": "^1.4.2",
     "through2": "^2.0.0",
-    "vinyl-paths": "^1.0.0"
+    "vinyl-paths": "^2.0.0"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=0.12.0"
   },
   "scripts": {
     "test": "gulp && git status | grep 'working directory clean' >/dev/null || (echo 'Please commit all changes generated by building'; exit 1)"
+  },
+  "babel": {
+    "only": "gulpfile.babel.js"
   }
 }
