@@ -17,29 +17,34 @@
  *
  */
 
+ // jscs:disable jsDoc
+
 'use strict';
 
 // Include Gulp & Tools We'll Use
-var gulp = require('gulp');
-var fs = require('fs');
-var merge = require('merge-stream');
-var $ = require('gulp-load-plugins')();
-var uniffe = require('./utils/uniffe.js');
-var del = require('del');
-var vinylPaths = require('vinyl-paths');
-var runSequence = require('run-sequence');
-var browserSync = require('browser-sync');
-var codeFiles = '';
-var reload = browserSync.reload;
-var path = require('path');
-var pkg = require('./package.json');
-var through = require('through2');
-var swig = require('swig');
-var hostedLibsUrlPrefix = 'https://storage.googleapis.com/code.getmdl.io';
-var bucketProd = 'gs://www.getmdl.io';
-var bucketStaging = 'gs://mdl-staging';
-var bucketCode = 'gs://code.getmdl.io';
-var banner = ['/**',
+import fs from 'fs';
+import path from 'path';
+import mergeStream from 'merge-stream';
+import del from 'del';
+import vinylPaths from'vinyl-paths';
+import runSequence from 'run-sequence';
+import browserSync from 'browser-sync';
+import through from 'through2';
+import swig from 'swig';
+import gulp from 'gulp';
+import closureCompiler from 'gulp-closure-compiler';
+import gulpLoadPlugins from 'gulp-load-plugins';
+import uniffe from './utils/uniffe.js';
+import pkg from './package.json';
+
+const $ = gulpLoadPlugins();
+const reload = browserSync.reload;
+const hostedLibsUrlPrefix = 'https://code.getmdl.io';
+const templateArchivePrefix = 'mdl-template-';
+const bucketProd = 'gs://www.getmdl.io';
+const bucketStaging = 'gs://mdl-staging';
+const bucketCode = 'gs://code.getmdl.io';
+const banner = ['/**',
   ' * <%= pkg.name %> - <%= pkg.description %>',
   ' * @version v<%= pkg.version %>',
   ' * @license <%= pkg.license %>',
@@ -48,7 +53,9 @@ var banner = ['/**',
   ' */',
   ''].join('\n');
 
-var AUTOPREFIXER_BROWSERS = [
+let codeFiles = '';
+
+const AUTOPREFIXER_BROWSERS = [
   'ie >= 10',
   'ie_mob >= 10',
   'ff >= 30',
@@ -60,31 +67,54 @@ var AUTOPREFIXER_BROWSERS = [
   'bb >= 10'
 ];
 
+const SOURCES = [
+  // Component handler
+  'src/mdlComponentHandler.js',
+  // Polyfills/dependencies
+  'src/third_party/**/*.js',
+  // Base components
+  'src/button/button.js',
+  'src/checkbox/checkbox.js',
+  'src/icon-toggle/icon-toggle.js',
+  'src/menu/menu.js',
+  'src/progress/progress.js',
+  'src/radio/radio.js',
+  'src/slider/slider.js',
+  'src/snackbar/snackbar.js',
+  'src/spinner/spinner.js',
+  'src/switch/switch.js',
+  'src/tabs/tabs.js',
+  'src/textfield/textfield.js',
+  'src/tooltip/tooltip.js',
+  // Complex components (which reuse base components)
+  'src/layout/layout.js',
+  'src/data-table/data-table.js',
+  // And finally, the ripples
+  'src/ripple/ripple.js'
+];
+
 // ***** Development tasks ****** //
 
 // Lint JavaScript
-gulp.task('jshint', function() {
-  return gulp.src(['src/**/*.js' , 'gulpfile.js'])
+gulp.task('lint', () => {
+  return gulp.src([
+      'src/**/*.js',
+      'gulpfile.babel.js'
+    ])
     .pipe(reload({stream: true, once: true}))
     .pipe($.jshint())
-    .pipe($.jshint.reporter('jshint-stylish'))
-    .pipe($.if(!browserSync.active, $.jshint.reporter('fail')));
-});
-
-// Lint JavaScript code style
-gulp.task('jscs', function() {
-  return gulp.src(['src/**/*.js' , 'gulpfile.js'])
-    .pipe(reload({stream: true, once: true}))
     .pipe($.jscs())
+    .pipe($.jshint.reporter('jshint-stylish'))
     .pipe($.jscs.reporter())
-    .pipe($.if(!browserSync.active, $.jshint.reporter('fail')));
+    .pipe($.if(!browserSync.active, $.jshint.reporter('fail')))
+    .pipe($.if(!browserSync.active, $.jscs.reporter('fail')));
 });
 
 // ***** Production build tasks ****** //
 
 // Optimize Images
 // TODO: Update image paths in final CSS to match root/images
-gulp.task('images', function() {
+gulp.task('images', () => {
   return gulp.src('src/**/*.{svg,png,jpg}')
     .pipe($.flatten())
     .pipe($.cache($.imagemin({
@@ -96,10 +126,8 @@ gulp.task('images', function() {
 });
 
 // Compile and Automatically Prefix Stylesheets (dev)
-gulp.task('styles:dev', function() {
-  return gulp.src([
-    'src/**/*.scss'
-  ])
+gulp.task('styles:dev', () => {
+  return gulp.src('src/**/*.scss')
     .pipe($.sass({
       precision: 10,
       onError: console.error.bind(console, 'Sass error:')
@@ -113,67 +141,59 @@ gulp.task('styles:dev', function() {
 });
 
 // Compile and Automatically Prefix Stylesheet Templates (production)
-gulp.task('styletemplates', function() {
+gulp.task('styletemplates', () => {
   // For best performance, don't add Sass partials to `gulp.src`
-  return gulp.src([
-    'src/template.scss'
-  ])
+  return gulp.src('src/template.scss')
     // Generate Source Maps
     .pipe($.sourcemaps.init())
     .pipe($.sass({
       precision: 10,
       onError: console.error.bind(console, 'Sass error:')
     }))
-    .pipe($.cssInlineImages({
-      webRoot: 'src'
-    }))
+    .pipe($.cssInlineImages({webRoot: 'src'}))
     .pipe($.autoprefixer(AUTOPREFIXER_BROWSERS))
     .pipe(gulp.dest('.tmp'))
     // Concatenate Styles
     .pipe($.concat('material.css.template'))
-    .pipe(gulp.dest('./dist'))
+    .pipe(gulp.dest('dist'))
     // Minify Styles
     .pipe($.if('*.css.template', $.csso()))
     .pipe($.concat('material.min.css.template'))
-    .pipe($.header(banner, {pkg: pkg}))
-    .pipe($.sourcemaps.write('./'))
-    .pipe(gulp.dest('./dist'))
+    .pipe($.header(banner, {pkg}))
+    .pipe($.sourcemaps.write('.'))
+    .pipe(gulp.dest('dist'))
     .pipe($.size({title: 'styles'}));
 });
 
 // Compile and Automatically Prefix Stylesheets (production)
-gulp.task('styles', function() {
+gulp.task('styles', () => {
   // For best performance, don't add Sass partials to `gulp.src`
-  return gulp.src([
-    'src/material-design-lite.scss'
-  ])
+  return gulp.src('src/material-design-lite.scss')
     // Generate Source Maps
     .pipe($.sourcemaps.init())
     .pipe($.sass({
       precision: 10,
       onError: console.error.bind(console, 'Sass error:')
     }))
-    .pipe($.cssInlineImages({
-      webRoot: 'src'
-    }))
+    .pipe($.cssInlineImages({webRoot: 'src'}))
     .pipe($.autoprefixer(AUTOPREFIXER_BROWSERS))
     .pipe(gulp.dest('.tmp'))
     // Concatenate Styles
     .pipe($.concat('material.css'))
-    .pipe($.header(banner, {pkg: pkg}))
-    .pipe(gulp.dest('./dist'))
+    .pipe($.header(banner, {pkg}))
+    .pipe(gulp.dest('dist'))
     // Minify Styles
     .pipe($.if('*.css', $.csso()))
     .pipe($.concat('material.min.css'))
-    .pipe($.header(banner, {pkg: pkg}))
-    .pipe($.sourcemaps.write('./'))
-    .pipe(gulp.dest('./dist'))
+    .pipe($.header(banner, {pkg}))
+    .pipe($.sourcemaps.write('.'))
+    .pipe(gulp.dest('dist'))
     .pipe($.size({title: 'styles'}));
 });
 
 // Only generate CSS styles for the MDL grid
-gulp.task('styles-grid', function() {
-  return gulp.src(['src/material-design-lite-grid.scss'])
+gulp.task('styles-grid', () => {
+  return gulp.src('src/material-design-lite-grid.scss')
     .pipe($.sass({
       precision: 10,
       onError: console.error.bind(console, 'Sass error:')
@@ -182,109 +202,123 @@ gulp.task('styles-grid', function() {
     .pipe(gulp.dest('.tmp'))
     // Concatenate Styles
     .pipe($.concat('material-grid.css'))
-    .pipe($.header(banner, {pkg: pkg}))
-    .pipe(gulp.dest('./dist'))
+    .pipe($.header(banner, {pkg}))
+    .pipe(gulp.dest('dist'))
     // Minify Styles
     .pipe($.if('*.css', $.csso()))
     .pipe($.concat('material-grid.min.css'))
-    .pipe($.header(banner, {pkg: pkg}))
-    .pipe(gulp.dest('./dist'))
+    .pipe($.header(banner, {pkg}))
+    .pipe(gulp.dest('dist'))
     .pipe($.size({title: 'styles-grid'}));
 });
 
+// Build with Google's Closure Compiler, requires Java 1.7+ installed.
+gulp.task('closure', () => {
+  return gulp.src(SOURCES)
+    .pipe(closureCompiler({
+      compilerPath: 'node_modules/google-closure-compiler/compiler.jar',
+      fileName: 'material.closure.min.js',
+      compilerFlags: {
+        // jscs:disable closureCamelCase
+        compilation_level: 'ADVANCED_OPTIMIZATIONS',
+        language_in: 'ECMASCRIPT6_STRICT',
+        language_out: 'ECMASCRIPT5_STRICT',
+        warning_level: 'VERBOSE'
+        // jscs:enable closureCamelCase
+      }
+    }))
+    .pipe(gulp.dest('./dist'));
+});
+
 // Concatenate And Minify JavaScript
-gulp.task('scripts', ['jscs', 'jshint'], function() {
-  var sources = [
-    // Component handler
-    'src/mdlComponentHandler.js',
-    // Polyfills/dependencies
-    'src/third_party/**/*.js',
-    // Base components
-    'src/button/button.js',
-    'src/checkbox/checkbox.js',
-    'src/icon-toggle/icon-toggle.js',
-    'src/menu/menu.js',
-    'src/progress/progress.js',
-    'src/radio/radio.js',
-    'src/slider/slider.js',
-    'src/spinner/spinner.js',
-    'src/switch/switch.js',
-    'src/tabs/tabs.js',
-    'src/textfield/textfield.js',
-    'src/tooltip/tooltip.js',
-    // Complex components (which reuse base components)
-    'src/layout/layout.js',
-    'src/data-table/data-table.js',
-    // And finally, the ripples
-    'src/ripple/ripple.js'
-  ];
-  return gulp.src(sources)
+gulp.task('scripts', ['lint'], () => {
+  return gulp.src(SOURCES)
     .pipe($.if(/mdlComponentHandler\.js/, $.util.noop(), uniffe()))
     .pipe($.sourcemaps.init())
     // Concatenate Scripts
     .pipe($.concat('material.js'))
-    .pipe($.iife({
-      useStrict: true,
-    }))
-    .pipe(gulp.dest('./dist'))
+    .pipe($.iife({useStrict: true}))
+    .pipe(gulp.dest('dist'))
     // Minify Scripts
     .pipe($.uglify({
       sourceRoot: '.',
       sourceMapIncludeSources: true
     }))
-    .pipe($.header(banner, {pkg: pkg}))
+    .pipe($.header(banner, {pkg}))
     .pipe($.concat('material.min.js'))
     // Write Source Maps
-    .pipe($.sourcemaps.write('./'))
-    .pipe(gulp.dest('./dist'))
+    .pipe($.sourcemaps.write('.'))
+    .pipe(gulp.dest('dist'))
     .pipe($.size({title: 'scripts'}));
 });
 
 // Clean Output Directory
-gulp.task('clean', del.bind(null, ['dist', '.publish'], {dot: true}));
+gulp.task('clean', () => del(['dist', '.publish']));
 
 // Copy package manger and LICENSE files to dist
-gulp.task('metadata', function() {
-  return gulp.src(['package.json', 'bower.json', 'LICENSE'])
-    .pipe(gulp.dest('./dist'));
+gulp.task('metadata', () => {
+  return gulp.src([
+      'package.json',
+      'bower.json',
+      'LICENSE'
+    ])
+    .pipe(gulp.dest('dist'));
 });
 
 // Build Production Files, the Default Task
-gulp.task('default', ['clean', 'mocha'], function(cb) {
+gulp.task('default', ['clean'], cb => {
   runSequence(
     ['styles', 'styles-grid'],
     ['scripts'],
+    ['mocha'],
     cb);
 });
 
 // Build production files and microsite
-gulp.task('all', ['clean', 'mocha'], function(cb) {
+gulp.task('all', ['clean'], cb => {
   runSequence(
-    ['default', 'styletemplates'],
-    ['styles:gen'],
-    ['jshint', 'jscs', 'scripts',  'assets', 'demos', 'pages',
-     'templates', 'images', 'styles-grid', 'metadata'],
+    ['styletemplates'],
+    ['styles-grid', 'styles:gen'],
+    ['scripts'],
+    ['mocha'],
+    ['assets', 'pages',
+     'templates', 'images', 'metadata'],
     ['zip'],
     cb);
 });
 
 // ***** Testing tasks ***** //
 
-gulp.task('mocha', ['styles'], function() {
-  return gulp.src('./test/index.html')
+gulp.task('mocha', ['styles'], () => {
+  return gulp.src('test/index.html')
     .pipe($.mochaPhantomjs({reporter: 'tap'}));
 });
 
-gulp.task('test', ['jshint', 'jscs', 'mocha']);
+gulp.task('mocha:closure', ['closure'], () => {
+  return gulp.src('test/index.html')
+    .pipe($.replace('src="../dist/material.js"',
+        'src="../dist/material.closure.min.js"'))
+    .pipe($.rename('temp.html'))
+    .pipe(gulp.dest('test'))
+    .pipe($.mochaPhantomjs({reporter: 'tap'}))
+    .on('finish', () => del.sync('test/temp.html'))
+    .on('error', () => del.sync('test/temp.html'));
+});
 
-gulp.task('test:visual', function() {
+gulp.task('test', [
+  'lint',
+  'mocha',
+  'mocha:closure'
+]);
+
+gulp.task('test:visual', () => {
   browserSync({
     notify: false,
-    server: './',
+    server: '.',
     startPath: 'test/visual/index.html'
   });
 
-  gulp.watch(['test/visual/**'], reload);
+  gulp.watch('test/visual/**', reload);
 });
 
 // ***** Landing page tasks ***** //
@@ -293,125 +327,127 @@ gulp.task('test:visual', function() {
  * Site metadata for use with templates.
  * @type {Object}
  */
-var site = {};
+const site = {};
 
 /**
  * Generates an HTML file based on a template and file metadata.
  */
 function applyTemplate() {
-  return through.obj(function(file, enc, cb) {
-    var data = {
-      site: site,
+  return through.obj((file, enc, cb) => {
+    const data = {
+      site,
       page: file.page,
       content: file.contents.toString()
     };
 
-    var templateFile = path.join(
-        __dirname, 'docs', '_templates', file.page.layout + '.html');
-    var tpl = swig.compileFile(templateFile, {cache: false});
-    file.contents = new Buffer(tpl(data), 'utf8');
-    this.push(file);
-    cb();
+    const templateFile = path.join(
+        __dirname, 'docs', '_templates', `${file.page.layout}.html`);
+    const tpl = swig.compileFile(templateFile, {cache: false});
+
+    file.contents = new Buffer(tpl(data));
+    cb(null, file);
   });
 }
 
 /**
  * Generates an index.html file for each README in MDL/src directory.
  */
-gulp.task('components', ['demos'], function() {
-  return gulp.src(['./src/**/README.md'], {base: './src'})
+gulp.task('components', ['demos'], () => {
+  return gulp.src('src/**/README.md', {base: 'src'})
     // Add basic front matter.
     .pipe($.header('---\nlayout: component\nbodyclass: component\ninclude_prefix: ../../\n---\n\n'))
-    .pipe($.frontMatter({property: 'page', remove: true}))
+    .pipe($.frontMatter({
+      property: 'page',
+      remove: true
+    }))
     .pipe($.marked())
-    .pipe((function() {
-      return through.obj(function(file, enc, cb) {
+    .pipe((() => {
+      return through.obj((file, enc, cb) => {
         file.page.component = file.relative.split('/')[0];
-        this.push(file);
-        cb();
+        cb(null, file);
       });
     })())
     .pipe(applyTemplate())
-    .pipe($.rename(function(path) {
-      path.basename = 'index';
-    }))
+    .pipe($.rename(path => path.basename = 'index'))
     .pipe(gulp.dest('dist/components'));
 });
 
 /**
  * Copies demo files from MDL/src directory.
  */
-gulp.task('demoresources', function() {
+gulp.task('demoresources', () => {
   return gulp.src([
-      './src/**/demos.css',
-      './src/**/demo.css',
-      './src/**/demo.js'
-    ], {base: './src'})
-      .pipe($.if('*.scss', $.sass({
-        precision: 10,
-        onError: console.error.bind(console, 'Sass error:')
-      })))
-      .pipe($.cssInlineImages({
-        webRoot: 'src'
-      }))
-      .pipe($.if('*.css', $.autoprefixer(AUTOPREFIXER_BROWSERS)))
-      .pipe(gulp.dest('dist/components'));
+      'src/**/demos.css',
+      'src/**/demo.css',
+      'src/**/demo.js'
+    ], {base: 'src'})
+    .pipe($.if('*.scss', $.sass({
+      precision: 10,
+      onError: console.error.bind(console, 'Sass error:')
+    })))
+    .pipe($.cssInlineImages({webRoot: 'src'}))
+    .pipe($.if('*.css', $.autoprefixer(AUTOPREFIXER_BROWSERS)))
+    .pipe(gulp.dest('dist/components'));
 });
 
 /**
  * Generates demo files for testing made of all the snippets and the demo file
  * put together.
  */
-gulp.task('demos', ['demoresources'], function() {
+gulp.task('demos', ['demoresources'], () => {
   /**
    * Retrieves the list of component folders.
    */
   function getComponentFolders() {
-    return fs.readdirSync('./src/')
-      .filter(function(file) {
-        return fs.statSync(path.join('./src/', file)).isDirectory();
-      });
+    return fs.readdirSync('src')
+      .filter(file => fs.statSync(path.join('src', file)).isDirectory());
   }
 
-  var tasks = getComponentFolders().map(function(component) {
+  const tasks = getComponentFolders().map(component => {
     return gulp.src([
-        './src/' + component + '/snippets/*.html',
-        './src/' + component + '/demo.html'
+        path.join('src', component, 'snippets', '*.html'),
+        path.join('src', component, 'demo.html')
       ])
       .pipe($.concat('/demo.html'))
       // Add basic front matter.
       .pipe($.header('---\nlayout: demo\nbodyclass: demo\ninclude_prefix: ../../\n---\n\n'))
-      .pipe($.frontMatter({property: 'page', remove: true}))
+      .pipe($.frontMatter({
+        property: 'page',
+        remove: true
+      }))
       .pipe($.marked())
-      .pipe((function() {
-        return through.obj(function(file, enc, cb) {
-            file.page.component = component;
-            this.push(file);
-            cb();
-          });
+      .pipe((() => {
+        return through.obj((file, enc, cb) => {
+          file.page.component = component;
+          cb(null, file);
+        });
       })())
       .pipe(applyTemplate())
-      .pipe(gulp.dest('dist/components/' + component));
+      .pipe(gulp.dest(path.join('dist', 'components', component)));
   });
 
-  return merge(tasks);
+  return mergeStream(tasks);
 });
 
 /**
  * Generates an HTML file for each md file in _pages directory.
  */
-gulp.task('pages', ['components'], function() {
-  return gulp.src(['docs/_pages/*.md'])
-    .pipe($.frontMatter({property: 'page', remove: true}))
+gulp.task('pages', ['components'], () => {
+  return gulp.src('docs/_pages/*.md')
+    .pipe($.frontMatter({
+      property: 'page',
+      remove: true
+    }))
     .pipe($.marked())
     .pipe(applyTemplate())
     .pipe($.replace('$$version$$', pkg.version))
     .pipe($.replace('$$hosted_libs_prefix$$', hostedLibsUrlPrefix))
+    .pipe($.replace('$$template_archive_prefix$$', templateArchivePrefix))
     /* Replacing code blocks class name to match Prism's. */
     .pipe($.replace('class="lang-', 'class="language-'))
     /* Translate html code blocks to "markup" because that's what Prism uses. */
     .pipe($.replace('class="language-html', 'class="language-markup'))
-    .pipe($.rename(function(path) {
+    .pipe($.rename(path => {
       if (path.basename !== 'index') {
         path.dirname = path.basename;
         path.basename = 'index';
@@ -423,7 +459,7 @@ gulp.task('pages', ['components'], function() {
 /**
  * Copies assets from MDL and _assets directory.
  */
-gulp.task('assets', function() {
+gulp.task('assets', () => {
   return gulp.src([
       'docs/_assets/**/*',
       'node_modules/clippy/build/clippy.swf',
@@ -443,8 +479,11 @@ gulp.task('assets', function() {
     })))
     .pipe($.if(/\.css/i, $.autoprefixer(AUTOPREFIXER_BROWSERS)))
     .pipe($.if(/\.css/i, $.csso()))
-    .pipe($.if(/\.js/i, $.uglify({preserveComments: 'some', sourceRoot: '.',
-      sourceMapIncludeSources: true})))
+    .pipe($.if(/\.js/i, $.uglify({
+      preserveComments: 'some',
+      sourceRoot: '.',
+      sourceMapIncludeSources: true
+    })))
     .pipe(gulp.dest('dist/assets'));
 });
 
@@ -467,7 +506,7 @@ function watch() {
 /**
  * Serves the landing page from "out" directory.
  */
-gulp.task('serve:browsersync', function() {
+gulp.task('serve:browsersync', () => {
   browserSync({
     notify: false,
     server: {
@@ -478,7 +517,7 @@ gulp.task('serve:browsersync', function() {
   watch();
 });
 
-gulp.task('serve', function() {
+gulp.task('serve', () => {
   $.connect.server({
     root: 'dist',
     port: 5000,
@@ -487,51 +526,65 @@ gulp.task('serve', function() {
 
   watch();
 
-  gulp.src('./dist/index.html')
-    .pipe($.open('', {url: 'http://localhost:5000'}));
+  gulp.src('dist/index.html')
+    .pipe($.open({uri: 'http://localhost:5000'}));
 });
 
 // Generate release archive containing just JS, CSS, Source Map deps
-gulp.task('zip:mdl', function() {
-  return gulp.src(['dist/material?(.min)@(.js|.css)?(.map)', 'LICENSE', 'bower.json', 'package.json'])
+gulp.task('zip:mdl', () => {
+  return gulp.src([
+      'dist/material?(.min)@(.js|.css)?(.map)',
+      'LICENSE',
+      'bower.json',
+      'package.json'
+    ])
     .pipe($.zip('mdl.zip'))
     .pipe(gulp.dest('dist'));
 });
 
-// Generate release archive containing the library, templates and assets
-// for templates. Note that it is intentional for some templates to include
-// a customised version of the material.min.css file for their own needs.
-// Others (e.g the Android template) simply use the default built version of
-// the library.
+/**
+ * Returns the list of children directories inside the given directory.
+ * @param {string} dir the parent directory
+ * @return {Array<string>} list of child directories
+ */
+function getSubDirectories(dir) {
+  return fs.readdirSync(dir)
+    .filter(file => fs.statSync(path.join(dir, file)).isDirectory());
+}
 
-// Define a filter containing only the build assets we want to pluck from the
-// `dist` stream. This enables us to preserve the correct final dir structure,
-// which was not occurring when simply using `gulp.src` in `zip:templates`
+// Generate release archives containing the templates and assets for templates.
+gulp.task('zip:templates', () => {
+  const templates = getSubDirectories('dist/templates');
 
-var fileFilter = $.filter([
-  'material?(.min)@(.js|.css)?(.map)',
-  'templates/**/*.*',
-  'assets/**/*.*',
-  'LICENSE',
-  'bower.json',
-  'package.json']);
+  // Generate a zip file for each template.
+  const generateZips = templates.map(template => {
+    return gulp.src([
+        `dist/templates/${template}/**/*.*`,
+        'LICENSE'
+      ])
+      .pipe($.rename(path => {
+        path.dirname = path.dirname.replace(`dist/templates/${template}`, '');
+      }))
+      .pipe($.zip(`${templateArchivePrefix}${template}.zip`))
+      .pipe(gulp.dest('dist'));
+  });
 
-gulp.task('zip:templates', function() {
-  // Stream of all `dist` files and other package manager files from root
-  return gulp.src(['dist/**/*.*', 'LICENSE', 'bower.json', 'package.json'])
-  .pipe(fileFilter)
-  .pipe($.zip('mdl-templates.zip'))
-  .pipe(fileFilter.restore())
-  .pipe(gulp.dest('dist'));
+  return mergeStream(generateZips);
 });
 
-gulp.task('zip', ['zip:templates', 'zip:mdl']);
+gulp.task('zip', [
+  'zip:templates',
+  'zip:mdl'
+]);
 
-gulp.task('genCodeFiles', function() {
-  return gulp.src(['dist/material.*@(js|css)?(.map)', 'dist/mdl.zip', 'dist/mdl-templates.zip'],
-      {read: false})
-    .pipe($.tap(function(file, t) {
-      codeFiles += ' dist/' + path.basename(file.path);
+gulp.task('genCodeFiles', () => {
+  return gulp.src([
+      'dist/material.*@(js|css)?(.map)',
+      'dist/mdl.zip',
+      `dist/${templateArchivePrefix}*.zip`
+    ], {read: false})
+    .pipe($.tap(file => {
+      codeFiles += ` dist/${path.basename(file.path)}`;
     }));
 });
 
@@ -540,9 +593,9 @@ gulp.task('genCodeFiles', function() {
 // global, high performance caching/content delivery network (CDN) service.
 // This task requires gsutil to be installed and configured.
 // For info on gsutil: https://cloud.google.com/storage/docs/gsutil.
-gulp.task('pushCodeFiles', function() {
-  var dest = bucketCode;
-  process.stdout.write('Publishing ' + pkg.version + ' to CDN (' + dest + ')\n');
+gulp.task('pushCodeFiles', () => {
+  const dest = bucketCode;
+  console.log(`Publishing ${pkg.version} to CDN (${dest})`);
 
   // Build cache control and gsutil cmd to copy
   // each object into a GCS bucket. The dest is a version specific path.
@@ -551,21 +604,21 @@ gulp.task('pushCodeFiles', function() {
   // (cache control, in this case).
   // Code files should NEVER be touched after uploading, therefore
   // 30 days caching is a safe value.
-  var cacheControl = '-h "Cache-Control:public,max-age=2592000"';
-  var gsutilCpCmd = 'gsutil -m cp -z js,css,map ';
-  var gsutilCacheCmd = 'gsutil -m setmeta -R ' + cacheControl;
+  const cacheControl = '-h "Cache-Control:public,max-age=2592000"';
+  const gsutilCpCmd = 'gsutil -m cp -z js,css,map ';
+  const gsutilCacheCmd = `gsutil -m setmeta -R ${cacheControl}`;
 
   // Upload the goodies to a separate GCS bucket with versioning.
   // Using a sep bucket avoids the risk of accidentally blowing away
   // old versions in the microsite bucket.
   return gulp.src('')
     .pipe($.shell([
-      gsutilCpCmd + codeFiles + ' ' + dest + '/' + pkg.version,
-      gsutilCacheCmd + ' ' + dest + '/' + pkg.version
+      `${gsutilCpCmd}${codeFiles} ${dest}/${pkg.version}`,
+      `${gsutilCacheCmd} ${dest}/${pkg.version}`
     ]));
 });
 
-gulp.task('publish:code', function(cb) {
+gulp.task('publish:code', cb => {
   runSequence(
     ['zip:mdl', 'zip:templates'],
     'genCodeFiles',
@@ -579,9 +632,10 @@ gulp.task('publish:code', function(cb) {
  * @param {string} pubScope the scope to publish to.
  */
 function mdlPublish(pubScope) {
-  var cacheTtl = null;
-  var src = null;
-  var dest = null;
+  let cacheTtl = null;
+  let src = null;
+  let dest = null;
+
   if (pubScope === 'staging') {
     // Set staging specific vars here.
     cacheTtl = 0;
@@ -595,23 +649,23 @@ function mdlPublish(pubScope) {
   } else if (pubScope === 'promote') {
     // Set promote (essentially prod) specific vars here.
     cacheTtl = 60;
-    src = bucketStaging + '/*';
+    src = `${bucketStaging}/*`;
     dest = bucketProd;
   }
 
-  var infoMsg = 'Publishing ' + pubScope + '/' + pkg.version + ' to GCS (' + dest + ')';
+  let infoMsg = `Publishing ${pubScope}/${pkg.version} to GCS (${dest})`;
   if (src) {
-    infoMsg += ' from ' + src;
+    infoMsg += ` from ${src}`;
   }
-  process.stdout.write(infoMsg + '\n');
+  console.log(infoMsg);
 
   // Build gsutil commands:
   // The gsutil -h option is used to set metadata headers.
   // The gsutil -m option requests parallel copies.
   // The gsutil -R option is used for recursive file copy.
-  var cacheControl = '-h "Cache-Control:public,max-age=' + cacheTtl + '"';
-  var gsutilCacheCmd = 'gsutil -m setmeta ' + cacheControl + ' ' + dest + '/**';
-  var gsutilCpCmd = 'gsutil -m cp -r -z html,css,js,svg ' + src + ' ' + dest;
+  const cacheControl = `-h "Cache-Control:public,max-age=${cacheTtl}"`;
+  const gsutilCacheCmd = `gsutil -m setmeta ${cacheControl} ${dest}/**`;
+  const gsutilCpCmd = `gsutil -m cp -r -z html,css,js,svg ${src} ${dest}`;
 
   gulp.src('').pipe($.shell([gsutilCpCmd, gsutilCacheCmd]));
 }
@@ -623,7 +677,7 @@ function mdlPublish(pubScope) {
 // This task requires gsutil to be installed and configured.
 // For info on gsutil: https://cloud.google.com/storage/docs/gsutil.
 //
-gulp.task('publish:prod', function() {
+gulp.task('publish:prod', () => {
   mdlPublish('prod');
 });
 
@@ -634,7 +688,7 @@ gulp.task('publish:prod', function() {
 // This task requires gsutil to be installed and configured.
 // For info on gsutil: https://cloud.google.com/storage/docs/gsutil.
 //
-gulp.task('publish:promote', function() {
+gulp.task('publish:promote', () => {
   mdlPublish('promote');
 });
 
@@ -644,47 +698,37 @@ gulp.task('publish:promote', function() {
 // This task requires gsutil to be installed and configured.
 // For info on gsutil: https://cloud.google.com/storage/docs/gsutil.
 //
-gulp.task('publish:staging', function() {
+gulp.task('publish:staging', () => {
   mdlPublish('staging');
 });
 
-gulp.task('templates:mdl', function() {
+gulp.task('_release', () => {
   return gulp.src([
-    'templates/**/*.scss'
-  ])
-    .pipe($.sass({
-      precision: 10,
-      onError: console.error.bind(console, 'Sass error:')
-    }))
-    .pipe($.cssInlineImages({
-      webRoot: 'src'
-    }))
-    .pipe($.autoprefixer(AUTOPREFIXER_BROWSERS))
-    .pipe($.csso())
-    .pipe($.rename({suffix: '.min'}))
-    .pipe(gulp.dest('dist/templates'));
-});
-
-gulp.task('_release', function() {
-  return gulp.src(['dist/material?(.min)@(.js|.css)?(.map)', 'LICENSE',
-    'README.md', 'bower.json', 'package.json', '.jscsrc', '.jshintrc',
-    './sr?/**/*', 'gulpfile.js', './util?/**/*'])
+      'dist/material?(.min)@(.js|.css)?(.map)',
+      'LICENSE',
+      'README.md',
+      'bower.json',
+      'package.json',
+      '.jscsrc',
+      '.jshintrc',
+      './sr?/**/*',
+      'gulpfile.babel.js',
+      './util?/**/*'
+    ])
     .pipe(gulp.dest('_release'));
 });
 
-gulp.task('publish:release', ['_release'], function() {
+gulp.task('publish:release', ['_release'], () => {
   return gulp.src('_release')
-  .pipe($.subtree({
-    remote: 'origin',
-    branch: 'release'
-  }))
-  .pipe(vinylPaths(del));
+    .pipe($.subtree({
+      remote: 'origin',
+      branch: 'release'
+    }))
+    .pipe(vinylPaths(del));
 });
 
-gulp.task('templates:styles', function() {
-  return gulp.src([
-    'templates/**/*.css'
-  ])
+gulp.task('templates:styles', () => {
+  return gulp.src('templates/**/*.css')
     .pipe($.autoprefixer(AUTOPREFIXER_BROWSERS))
     // FIXME: This crashes. It's a bug in gulp-csso,
     // not csso itself.
@@ -692,58 +736,72 @@ gulp.task('templates:styles', function() {
     .pipe(gulp.dest('dist/templates'));
 });
 
-gulp.task('templates:static', function() {
-  return gulp.src([
-    'templates/**/*.html',
-  ])
+gulp.task('templates:static', () => {
+  return gulp.src('templates/**/*.html')
+  .pipe($.replace('$$version$$', pkg.version))
+  .pipe($.replace('$$hosted_libs_prefix$$', hostedLibsUrlPrefix))
   .pipe(gulp.dest('dist/templates'));
 });
 
-gulp.task('templates:images', function() {
-  return gulp.src([
-    'templates/*/images/**/*'
-  ])
-  .pipe($.imagemin({
-    progressive: true,
-    interlaced: true
-  }))
-  .pipe(gulp.dest('dist/templates'));
+// This task can be used if you want to test the templates against locally
+// built version of the MDL libraries.
+gulp.task('templates:localtestingoverride', () => {
+  return gulp.src('templates/**/*.html')
+    .pipe($.replace('$$version$$', '.'))
+    .pipe($.replace('$$hosted_libs_prefix$$', ''))
+    .pipe(gulp.dest('dist/templates'));
 });
 
-gulp.task('templates:fonts', function() {
-  return gulp.src([
-    'templates/*/fonts/**/*'
-  ])
-  .pipe(gulp.dest('dist/templates/'));
+gulp.task('templates:images', () => {
+  return gulp.src('templates/*/images/**/*')
+    .pipe($.imagemin({
+      progressive: true,
+      interlaced: true
+    }))
+    .pipe(gulp.dest('dist/templates'));
 });
 
-gulp.task('templates', ['templates:static', 'templates:images', 'templates:mdl',
-    'templates:fonts', 'templates:styles']);
+gulp.task('templates:fonts', () => {
+  return gulp.src('templates/*/fonts/**/*')
+    .pipe(gulp.dest('dist/templates/'));
+});
 
-gulp.task('styles:gen', ['styles'], function() {
-  var MaterialCustomizer = require('./docs/_assets/customizer.js');
-  var templatePath = path.join(__dirname, 'dist', 'material.min.css.template');
+gulp.task('templates', [
+  'templates:static',
+  'templates:images',
+  'templates:fonts',
+  'templates:styles'
+]);
+
+gulp.task('styles:gen', ['styles'], () => {
+  const MaterialCustomizer = require('./docs/_assets/customizer.js');
+  const templatePath = path.join(__dirname, 'dist', 'material.min.css.template');
   // TODO: This task needs refactoring once we turn MaterialCustomizer
   // into a proper Node module.
-  var mc = new MaterialCustomizer();
+  const mc = new MaterialCustomizer();
   mc.template = fs.readFileSync(templatePath).toString();
 
-  var stream = gulp.src('');
-  mc.paletteIndices.forEach(function(primary) {
-    mc.paletteIndices.forEach(function(accent) {
+  let stream = gulp.src('');
+
+  mc.paletteIndices.forEach(primary => {
+    mc.paletteIndices.forEach(accent => {
       if (primary === accent) {
         return;
       }
+
       if (mc.forbiddenAccents.indexOf(accent) !== -1) {
         return;
       }
-      var primaryName = primary.toLowerCase().replace(' ', '_');
-      var accentName = accent.toLowerCase().replace(' ', '_');
+
+      const primaryName = primary.toLowerCase().replace(' ', '_');
+      const accentName = accent.toLowerCase().replace(' ', '_');
+
       stream = stream.pipe($.file(
-        'material.' + primaryName + '-' + accentName + '.min.css',
+        `material.${primaryName}-${accentName}.min.css`,
         mc.processTemplate(primary, accent)
       ));
     });
   });
+
   stream.pipe(gulp.dest('dist'));
 });
