@@ -24,7 +24,7 @@
    * https://github.com/jasonmayes/mdl-component-design-pattern
    *
    * @constructor
-   * @param {HTMLElement} element The element that will be upgraded.
+   * @param {Element} element The element that will be upgraded.
    */
   var MaterialDataTable = function MaterialDataTable(element) {
     this.element_ = element;
@@ -66,7 +66,7 @@
    * single row (or multiple rows).
    *
    * @param {Element} checkbox Checkbox that toggles the selection state.
-   * @param {HTMLElement} row Row to toggle when checkbox changes.
+   * @param {Element} row Row to toggle when checkbox changes.
    * @param {(Array<Object>|NodeList)=} opt_rows Rows to toggle when checkbox changes.
    * @private
    */
@@ -106,7 +106,7 @@
    * Creates a checkbox for a single or or multiple rows and hooks up the
    * event handling.
    *
-   * @param {HTMLElement} row Row to toggle when checkbox changes.
+   * @param {Element} row Row to toggle when checkbox changes.
    * @param {(Array<Object>|NodeList)=} opt_rows Rows to toggle when checkbox changes.
    * @private
    */
@@ -122,8 +122,13 @@
     var checkbox = document.createElement('input');
     checkbox.type = 'checkbox';
     checkbox.classList.add('mdl-checkbox__input');
-    checkbox.addEventListener('change',
-        this.selectRow_(checkbox, row, opt_rows));
+
+    if (row) {
+      checkbox.checked = row.classList.contains(this.CssClasses_.IS_SELECTED);
+      checkbox.addEventListener('change', this.selectRow_(checkbox, row));
+    } else if (opt_rows) {
+      checkbox.addEventListener('change', this.selectRow_(checkbox, null, opt_rows));
+    }
 
     label.appendChild(checkbox);
     componentHandler.upgradeElement(label, 'MaterialCheckbox');
@@ -136,7 +141,9 @@
   MaterialDataTable.prototype.init = function() {
     if (this.element_) {
       var firstHeader = this.element_.querySelector('th');
-      var rows = this.element_.querySelector('tbody').querySelectorAll('tr');
+      var bodyRows = Array.prototype.slice.call(this.element_.querySelectorAll('tbody tr'));
+      var footRows = Array.prototype.slice.call(this.element_.querySelectorAll('tfoot tr'));
+      var rows = bodyRows.concat(footRows);
 
       if (this.element_.classList.contains(this.CssClasses_.SELECTABLE)) {
         var th = document.createElement('th');
@@ -148,14 +155,15 @@
           var firstCell = rows[i].querySelector('td');
           if (firstCell) {
             var td = document.createElement('td');
-            var rowCheckbox = this.createCheckbox_(rows[i]);
-            td.appendChild(rowCheckbox);
+            if (rows[i].parentNode.nodeName.toUpperCase() === 'TBODY') {
+              var rowCheckbox = this.createCheckbox_(rows[i]);
+              td.appendChild(rowCheckbox);
+            }
             rows[i].insertBefore(td, firstCell);
           }
         }
+        this.element_.classList.add(this.CssClasses_.IS_UPGRADED);
       }
-
-      this.element_.classList.add(this.CssClasses_.IS_UPGRADED);
     }
   };
 
