@@ -34,7 +34,8 @@ function uniffe(contents) {
 
   escodegen.attachComments(ast, comments, tokens);
 
-  if (ast.body[0].expression.callee === undefined) {
+  if (ast.body[0].expression === undefined ||
+      ast.body[0].expression.callee === undefined) {
     return contents;
   }
 
