@@ -28,8 +28,8 @@
 		'mdPickers'
 	])
 	
-	.constant('api', 'http://localhost/inventario/server/public/api/');
-	// .constant('api', 'http://api.medicavial.mx/api/');
+	// .constant('api', 'http://localhost/inventario/server/public/api/');
+	.constant('api', 'http://api.medicavial.mx/api/');
 
 })();
 
